+++ lib/src/events.dart
@@ -593,15 +593,15 @@ class RtcEngineEventHandler {
   /// - [int] `soundId`: ID of the local audio effect. Each local audio effect has a unique ID.
   SoundIdCallback audioEffectFinished;
 
-  /// Occurs when the state of the RTMP streaming changes.
+  /// Occurs when the state of the RTMP or RTMPS streaming changes.
   ///
   /// The SDK triggers this callback to report the result of the local user calling the [RtcEngine.addPublishStreamUrl] or [RtcEngine.removePublishStreamUrl] method. This callback returns the URL and its current streaming state. When the streaming state is [RtmpStreamingState.Failure], see the `errCode` parameter for details.
   ///
-  /// This callback indicates the state of the RTMP streaming. When exceptions occur, you can troubleshoot issues by referring to the detailed error descriptions in the `errCode` parameter.
+  /// This callback indicates the state of the RTMP or RTMPS streaming. When exceptions occur, you can troubleshoot issues by referring to the detailed error descriptions in the `errCode` parameter.
   ///
   /// The `RtmpStreamingStateCallback` typedef includes the following parameters:
-  /// - [String] `url`: The RTMP URL address.
-  /// - [RtmpStreamingState] `state`: The RTMP streaming state.
+  /// - [String] `url`: The CDN streaming URL.
+  /// - [RtmpStreamingState] `state`: The RTMP or RTMPS streaming state.
   /// - [RtmpStreamingErrorCode] `errCode`: The detailed error information for streaming.
   RtmpStreamingStateCallback rtmpStreamingStateChanged;
 
@@ -757,7 +757,7 @@ class RtcEngineEventHandler {
   /// This callback indicates whether you have successfully added an RTMP stream to the CDN.
   ///
   /// The `UrlWithErrorCallback` includes the following parameters:
-  /// - [String] `url`: The RTMP URL address.
+  /// - [String] `url`: The CDN streaming URL.
   /// - [ErrorCode] `error`: The detailed error information.
   @deprecated
   UrlWithErrorCallback streamPublished;
@@ -769,7 +769,7 @@ class RtcEngineEventHandler {
   /// This callback indicates whether you have successfully removed an RTMP stream from the CDN.
   ///
   /// The `UrlCallback` typedef includes the following parameter:
-  /// - [String] `url`: The RTMP URL address.
+  /// - [String] `url`: The CDN streaming URL.
   @deprecated
   UrlCallback streamUnpublished;
 
@@ -930,8 +930,6 @@ class RtcEngineEventHandler {
 
   /// Occurs when the first audio frame is published.
   ///
-  /// @since v3.1.2.
-  ///
   /// The SDK triggers this callback under one of the following circumstances:
   /// - The local client enables the audio module and calls [RtcEngine.joinChannel] successfully.
   /// - The local client calls [RtcEngine.muteLocalAudioStream] (`true`) and [RtcEngine.muteLocalAudioStream] (`false`) in sequence.
@@ -942,8 +940,6 @@ class RtcEngineEventHandler {
   ElapsedCallback firstLocalAudioFramePublished;
 
   /// Occurs when the first video frame is published.
-  ///
-  /// @since v3.1.2.
   ///
   /// The SDK triggers this callback under one of the following circumstances:
   /// - The local client enables the video module and calls [RtcEngine.joinChannel] successfully.
@@ -956,8 +952,6 @@ class RtcEngineEventHandler {
 
   /// Occurs when the audio publishing state changes.
   ///
-  /// @since v3.1.2.
-  ///
   /// This callback indicates the publishing state change of the local audio stream.
   ///
   /// The `StreamPublishStateCallback` typedef includes the following parameters:
@@ -968,8 +962,6 @@ class RtcEngineEventHandler {
   StreamPublishStateCallback audioPublishStateChanged;
 
   /// Occurs when the video publishing state changes.
-  ///
-  /// @since v3.1.2.
   ///
   /// This callback indicates the publishing state change of the local video stream.
   ///
@@ -982,8 +974,6 @@ class RtcEngineEventHandler {
 
   /// Occurs when the audio subscribing state changes.
   ///
-  /// @since v3.1.2.
-  ///
   /// This callback indicates the subscribing state change of a remote audio stream.
   ///
   /// The `StreamSubscribeStateCallback` typedef includes the following parameters:
@@ -995,8 +985,6 @@ class RtcEngineEventHandler {
 
   /// Occurs when the video subscribing state changes.
   ///
-  /// @since v3.1.2.
-  ///
   /// This callback indicates the subscribing state change of a remote video stream.
   ///
   /// The `StreamSubscribeStateCallback` typedef includes the following parameters:
@@ -1006,12 +994,10 @@ class RtcEngineEventHandler {
   /// - [int] `elapseSinceLastState`: The time elapsed (ms) from the previous state to the current state.
   StreamSubscribeStateCallback videoSubscribeStateChanged;
 
-  /// Reports events during the RTMP streaming.
-  ///
-  /// @since v3.1.2.
+  /// Reports events during the RTMP or RTMPS streaming.
   ///
   /// The `RtmpStreamingEventCallback` typedef includes the following parameters:
-  /// - [String] `url`: The RTMP streaming URL.
+  /// - [String] `url`: The RTMP or RTMPS streaming URL.
   /// - [RtmpStreamingEvent] `eventCode`: The event code. See [RtmpStreamingEvent].
   RtmpStreamingEventCallback rtmpStreamingEvent;
 
@@ -1642,15 +1628,15 @@ class RtcChannelEventHandler {
   /// - [RemoteAudioStats] `stats`: Statistics of the received remote audio streams.
   RemoteAudioStatsCallback remoteAudioStats;
 
-  /// Occurs when the state of the RTMP streaming changes.
+  /// Occurs when the state of the RTMP or RTMPS streaming changes.
   ///
   /// The SDK triggers this callback to report the result of the local user calling the [RtcChannel.addPublishStreamUrl] or [RtcChannel.removePublishStreamUrl] method. This callback returns the URL and its current streaming state. When the streaming state is [RtmpStreamingState.Failure], see the `errCode` parameter for details.
   ///
-  /// This callback indicates the state of the RTMP streaming. When exceptions occur, you can troubleshoot issues by referring to the detailed error descriptions in the `errCode` parameter.
+  /// This callback indicates the state of the RTMP or RTMPS streaming. When exceptions occur, you can troubleshoot issues by referring to the detailed error descriptions in the `errCode` parameter.
   ///
   /// The `RtmpStreamingStateCallback` typedef includes the following parameters:
-  /// - [String] `url`: The RTMP URL address.
-  /// - [RtmpStreamingState] `state`: The RTMP streaming state.
+  /// - [String] `url`: The CDN streaming URL.
+  /// - [RtmpStreamingState] `state`: The RTMP or RTMPS streaming state.
   /// - [RtmpStreamingErrorCode] `errCode`: The detailed error information for streaming.
   RtmpStreamingStateCallback rtmpStreamingStateChanged;
 
@@ -1721,8 +1707,6 @@ class RtcChannelEventHandler {
 
   /// Occurs when the audio publishing state changes.
   ///
-  /// @since v3.1.2.
-  ///
   /// This callback indicates the publishing state change of the local audio stream.
   ///
   /// The `StreamPublishStateCallback` typedef includes the following parameters:
@@ -1733,8 +1717,6 @@ class RtcChannelEventHandler {
   StreamPublishStateCallback audioPublishStateChanged;
 
   /// Occurs when the video publishing state changes.
-  ///
-  /// @since v3.1.2.
   ///
   /// This callback indicates the publishing state change of the local video stream.
   ///
@@ -1747,8 +1729,6 @@ class RtcChannelEventHandler {
 
   /// Occurs when the audio subscribing state changes.
   ///
-  /// @since v3.1.2.
-  ///
   /// This callback indicates the subscribing state change of a remote audio stream.
   ///
   /// The `StreamSubscribeStateCallback` typedef includes the following parameters:
@@ -1760,8 +1740,6 @@ class RtcChannelEventHandler {
 
   /// Occurs when the video subscribing state changes.
   ///
-  /// @since v3.1.2.
-  ///
   /// This callback indicates the subscribing state change of a remote video stream.
   ///
   /// The `StreamSubscribeStateCallback` typedef includes the following parameters:
@@ -1771,12 +1749,10 @@ class RtcChannelEventHandler {
   /// - [int] `elapseSinceLastState`: The time elapsed (ms) from the previous state to the current state.
   StreamSubscribeStateCallback videoSubscribeStateChanged;
 
-  /// Reports events during the RTMP streaming.
-  ///
-  /// @since v3.1.2.
+  /// Reports events during the RTMP or RTMPS streaming.
   ///
   /// The `RtmpStreamingEventCallback` typedef includes the following parameters:
-  /// - [String] `url`: The RTMP streaming URL.
+  /// - [String] `url`: The RTMP or RTMPS streaming URL.
   /// - [RtmpStreamingEvent] `eventCode`: The event code. See [RtmpStreamingEvent].
   RtmpStreamingEventCallback rtmpStreamingEvent;