+++ test/agora_rtc_engine_test.dart
@@ -1,3 +1,5 @@
+import 'package:agora_rtc_engine/rtc_engine.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -10,6 +12,12 @@ void main() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
       return '42';
     });
+  });
+
+  test('LiveTranscoding', () {
+    var transcoding = LiveTranscoding(null, backgroundColor: Colors.red);
+    print(transcoding.backgroundColor);
+    print(transcoding.toJson());
   });
 
   tearDown(() {